Initial Framework commit. Includes functional helper and connection examples. These changes are not tested for full functionality and this framework is not in a release state.
Big todos include:
Authentication through HybridAuth, Format handler for basic formats
(JSON, XML, Serialized PHP, and HTML), and a modification on how
PageController is included as the base for all page requests. This is an
initial development commit so I can save changes to version control in
the event of a catastrophe.

Change-Id: I7c37056f8dece4d42d5d001f5c6558279220668c