Initial framework commit (not fully functional)
This commit has working helper and connection examples and is a part of
a larger effort to refactor code to use features from PHP 5.4 (traits)
and abstract logic for individual connection types out of the framework
base itself.

This source is not tested and is not expected to be adoption ready.