Added schema, models, and actions for example site.
Fixed a bug in MySQL's readRow call that was addressing the wrong data
variable and causing readRow not to function correctly for subsequent
queries.
Fixed the action calls for S3 to match the connector name.
Added layering to the Image helper and also added support for memory
limit increases through the ini.
Fixed Scaffolding for Actions and Models so they now take into account
more details for date and time and boolean fields and fixed the pathing.
Added functionality to the MySQL and Oracle time formats in
DateTimeObject.
Cleaned up options that haven't been implemented from the example.ini
file for Startup.
Added the SecureHash helper for generating secure salted hashes.
Added the schema and models and actions generated from it to the
example site code.