Changed how request handling worked for components. No longer support the option to pass the component name in the path.
The option to pass the component name in the path was taking precidence
over the component controller path which wasnt even being utilized,
however it was forcing a specification of a variable that in most cases
would never be used. The functionality can still be had for mapping
other components in a page controller, but this functionality would be
done by a switch case from a standard query string var rather than
through the sugar methods. This makes handling requests a 1 line
affair, and mapping output to different formats can essentially be the
same.