Added email confirmation table to db for registrations, and added sugar methods for session flashing to controllers.
Example of code inside a Controller accessing a session flash to show
off an error with bootstrap styling: $this->flashError('ErrorName',
'This is an error');

Same goes for confirmations and notifications. Other bootstrap
notification styles can be added by developers if they want, but out of
the box this supports three for simplicity sake.